+++ src/HealthcareApis/HealthcareApis/HealthcareApis/NewAzureRmHealthcareApisService.cs
@@ -161,6 +161,12 @@ namespace Microsoft.Azure.Commands.HealthcareApis.Commands
             HelpMessage = "Run cmdlet as a job in the background.")]
         public SwitchParameter AsJob { get; set; }
 
+        [Parameter(
+            Mandatory = false,
+            HelpMessage = "The network access type for Fhir Service. Commonly `Enabled` or `Disabled`.")]
+        [ValidateSet("Enabled", "Disabled", IgnoreCase = true)]
+        public string PublicNetworkAccess { get; set; }
+
         public override void ExecuteCmdlet()
         {
             try
@@ -183,13 +189,18 @@ namespace Microsoft.Azure.Commands.HealthcareApis.Commands
                             CosmosDbConfiguration = new ServiceCosmosDbConfigurationInfo() { OfferThroughput = GetCosmosDBThroughput(), KeyVaultKeyUri = GetCosmosDBKeyVaultKeyUri() },
                             CorsConfiguration = new ServiceCorsConfigurationInfo() { Origins = CorsOrigin, Headers = CorsHeader, Methods = CorsMethod, MaxAge = CorsMaxAge, AllowCredentials = AllowCorsCredential },
                             ExportConfiguration = new ServiceExportConfigurationInfo() { StorageAccountName = ExportStorageAccountName },
-                            AccessPolicies = accessPolicies
+                            AccessPolicies = accessPolicies,
                         }
                     };
 
                     if (this.ManagedIdentity.IsPresent)
                     {
-                        servicesDescription.Identity = new ServicesResourceIdentity() { Type = "SystemAssigned" };
+                        servicesDescription.Identity = new Management.HealthcareApis.Models.ResourceIdentity() { Type = "SystemAssigned" };
+                    }
+
+                    if (!string.IsNullOrEmpty(PublicNetworkAccess))
+                    {
+                        servicesDescription.Properties.PublicNetworkAccess = PublicNetworkAccess;
                     }
 
                     if (ShouldProcess(this.Name, Resources.createService))