+++ src/HealthcareApis/HealthcareApis/HealthcareApis/SetAzureRmHealthcareApisService.cs
@@ -253,9 +253,23 @@ namespace Microsoft.Azure.Commands.HealthcareApis.Commands
         public string ResourceId { get; set; }
 
         [Parameter(
-         Mandatory = false,
-         HelpMessage = "Run cmdlet as a job in the background.")]
+            Mandatory = false,
+            HelpMessage = "Run cmdlet as a job in the background.")]
         public SwitchParameter AsJob { get; set; }
+
+        [Parameter(
+            Mandatory = false,
+            ParameterSetName = ServiceNameParameterSet,
+            HelpMessage = "The network access type for Fhir service. Commonly `Enabled` or `Disabled`.")]
+        [Parameter(
+            Mandatory = false,
+            ParameterSetName = ResourceIdParameterSet,
+            HelpMessage = "The network access type for Fhir service. Commonly `Enabled` or `Disabled`.")]
+        [ValidateSet(
+            "Enabled",
+            "Disabled",
+            IgnoreCase = true)]
+        public string PublicNetworkAccess { get; set; }
 
         public override void ExecuteCmdlet()
         {
@@ -396,7 +410,8 @@ namespace Microsoft.Azure.Commands.HealthcareApis.Commands
                     {
                         StorageAccountName = ExportStorageAccountName ?? healthcareApisAccount.Properties.ExportConfiguration.StorageAccountName
                     },
-                    AccessPolicies = accessPolicies
+                    AccessPolicies = accessPolicies,
+                    PublicNetworkAccess = PublicNetworkAccess ?? healthcareApisAccount.Properties.PublicNetworkAccess
                 },
                 Kind = healthcareApisAccount.Kind,
                 Tags = GetTags(healthcareApisAccount)
@@ -404,7 +419,7 @@ namespace Microsoft.Azure.Commands.HealthcareApis.Commands
 
             if (this.EnableManagedIdentity.ToBool() && healthcareApisAccount.Identity == null)
             {
-                servicesDescription.Identity = new ServicesResourceIdentity() { Type = "SystemAssigned" };
+                servicesDescription.Identity = new Management.HealthcareApis.Models.ResourceIdentity() { Type = "SystemAssigned" };
             }
             else if (!this.DisableManagedIdentity.ToBool())
             {
@@ -412,7 +427,7 @@ namespace Microsoft.Azure.Commands.HealthcareApis.Commands
             }
             else
             {
-                servicesDescription.Identity = new ServicesResourceIdentity() { Type = "None" };
+                servicesDescription.Identity = new Management.HealthcareApis.Models.ResourceIdentity() { Type = "None" };
             }
 
             return servicesDescription;
@@ -498,7 +513,8 @@ namespace Microsoft.Azure.Commands.HealthcareApis.Commands
                     {
                         StorageAccountName = InputObject.ExportStorageAccountName ?? healthcareApisAccount.Properties.ExportConfiguration.StorageAccountName
                     },
-                    AccessPolicies = accessPolicies
+                    AccessPolicies = accessPolicies,
+                    PublicNetworkAccess = healthcareApisAccount.Properties.PublicNetworkAccess
                 },
                 Kind = ParseKind(InputObject.Kind),
                 Tags = InputObject.Tags
@@ -506,7 +522,7 @@ namespace Microsoft.Azure.Commands.HealthcareApis.Commands
 
             if (!String.IsNullOrEmpty(InputObject.IdentityType))
             {
-                servicesDescription.Identity = new ServicesResourceIdentity(InputObject.IdentityPrincipalId, InputObject.IdentityTenantId, InputObject.IdentityType);
+                servicesDescription.Identity = new Management.HealthcareApis.Models.ResourceIdentity(InputObject.IdentityPrincipalId, InputObject.IdentityTenantId, InputObject.IdentityType);
             }
             else
             {