+++ src/HealthcareApis/HealthcareApis/ChangeLog.md
@@ -21,6 +21,7 @@
 
 ## Version 3.0.0
 * Updated the SDK version to 3.0.0
+* Added support for Private Link
 * Added support for customer managed keys
 
 ## Version 1.1.0