+++ src/HealthcareApis/HealthcareApis/ChangeLog.md
@@ -18,11 +18,11 @@
         - Additional information about change #1
 -->
 ## Upcoming Release
+* Added support for customer managed keys
 
 ## Version 3.0.0
 * Updated the SDK version to 3.0.0
 * Added support for Private Link
-* Added support for customer managed keys
 
 ## Version 1.1.0
 * Updated the SDK version to 1.1.0