+++ src/HealthcareApis/HealthcareApis/ChangeLog.md
@@ -18,6 +18,8 @@
         - Additional information about change #1
 -->
 ## Upcoming Release
+
+## Version 1.2.0
 * Added support for customer managed keys
 
 ## Version 3.0.0